# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: HEI Admin API
  description: |
    _Programmatically connect to a computer programming [school](https://hei.school)._
    After [joining us](mailto:contact@mail.hei.school),
    you can get an identification token from
    our [dev](https://dev-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=5s8cg50doahmu855rlc8fr6qmp&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-dev.hei.school%2Fwhoami)
    or [prod](https://prod-hei-admin.auth.eu-west-3.amazoncognito.com/oauth2/authorize?client_id=i8bg538jpfu6mqmqb61m26trd&response_type=token&scope=email+openid&redirect_uri=https%3A%2F%2Fapi-prod.hei.school%2Fwhoami)
    authentication service.
    Then, start playing with our system!

    The implementation of our API is [publicly disclosed](https://github.com/hei-school/hei-admin-api).
    You are welcome to try and compromise it. Happy hacking!
servers:
  - url: 'https://api-dev.hei.school'
  - url: 'https://api-prod.hei.school'
security:
  - BearerAuth: [ ]
paths:
...
  /students/{student_id}/fees/{fee_id}:
    get:
      tags:
        - Paying
      summary: Get student fee by identifier
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
        - name: fee_id
          in: path
          required: true
          schema:
            type: string
      operationId: getStudentFeeById
      responses:
        '200':
          description: Identified student fee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Paying
      summary: Delete student fee by identifier.
      description: Soft delete
      operationId: deleteStudentFeeById
      parameters:
        - name: fee_id
          in: path
          required: true
          schema:
            type: string
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: student fee deleted by identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /students/{student_id}/fees:
    get:
      tags:
        - Paying
      summary: Get all student fees ordered by due datetime desc
      description: Fees can be filtered with an optional payment status parameter.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum:
              - UNPAID
              - PAID
              - LATE
      operationId: getStudentFees
      responses:
        '200':
          description: List of all student fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    post:
      tags:
        - Paying
      summary: Create student fees
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      operationId: createStudentFees
      requestBody:
        description: Student fees to create
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateFee'
      responses:
        '200':
          description: Created student fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Paying
      summary: Update student fees, can only modify due_datetime and comment, must contain an ID
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      operationId: updateStudentFees
      requestBody:
        description: Student fees to update
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Fee'
      responses:
        '200':
          description: Update student fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /fees/raw:
    get:
      tags:
        - Paying
      summary: Generate fees list as xlsx
      operationId: generateFeesListAsXlsx
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/FeeStatusEnum'
      responses:
        '200':
          description: List of fees in a xlsx file
          content:
            application/xlsx:
              schema:
                type: string
                format: byte
                example: XLSX VERSION
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /fees:
    put:
      tags:
        - Paying
      summary: Crupdate student fees
      operationId: crupdateStudentFees
      requestBody:
        description: Student fees to update if id is provided, create otherwise.
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateStudentFee'
      responses:
        '200':
          description: Created or updated student fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fee'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Paying
      summary: Get all student fees filtered by status and month
      parameters:
        - name: transaction_status
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/MpbsStatus'
        - name: type
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/FeeTypeEnum'
        - name: status
          in: query
          description: "See the PaymentStatus object for its value."
          schema:
            type: string
            default: LATE
            enum:
              - UNPAID
              - LATE
              - PAID
        - name: month_from
          in: query
          schema:
            type: string
            format: date-time
          required: false
        - name: month_to
          in: query
          schema:
            type: string
            format: date-time
          required: false
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: isMpbs
          in: query
          schema:
            type: boolean
        - name: student_ref
          in: query
          required: false
          schema:
            type: string
      operationId: getFees
      responses:
        '200':
          description: List of filtered student fees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeesWithStats'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /fees/stats:
    get:
      tags:
        - Paying
      summary: Get monthly statistics on Fees, the actual month by default
      parameters:
        - name: month_from
          in: query
          schema:
            type: string
            format: date-time
          required: false
        - name: month_to
          in: query
          schema:
            type: string
            format: date-time
          required: false
      operationId: getFeesStats
      responses:
        '200':
          description: Fees stats info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeesStatistics'
  /fees/templates:
    get:
      tags:
        - Paying
      summary: Get all fee templates
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: amount
          in: query
          schema:
            type: integer
        - name: number_of_payments
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
      operationId: getFeeTemplates
      responses:
        '200':
          description: List of templates of fee
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeeTemplate'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /fees/templates/{id}:
    get:
      tags:
        - Paying
      summary: Get fee template by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      description: Get fee template by `id` if `id` exists
      operationId: getFeeTemplateById
      responses:
        '200':
          description: The fee template requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeTemplate'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Paying
      summary: Create a new fee type or update existing fee type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      description: Update fee type if `id` exist, create it otherwise.
      operationId: crupdateFeeTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrupdateFeeTemplate'
      responses:
        '200':
          description: The created or updated fee type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeTemplate'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    CreateFee:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/FeeTypeEnum'
        comment:
          type: string
        total_amount:
          type: integer
        creation_datetime:
          type: string
          format: date-time
        due_datetime:
          type: string
          format: date-time
    CrupdateStudentFee:
      properties:
        id:
          type: string
        student_id:
          type: string
      allOf:
        - $ref: '#/components/schemas/CreateFee'
...
    FeesWithStats:
      properties:
        statistics:
          $ref: '#/components/schemas/FeesStatistics'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Fee'
    Fee:
      allOf:
        - $ref: '#/components/schemas/CreateFee'
      type: object
      properties:
        id:
          type: string
        student_ref:
          type: string
        student_id:
          type: string
        remaining_amount:
          type: integer
        status:
          $ref: '#/components/schemas/FeeStatusEnum'
        updated_at:
          type: string
          format: date-time
        mpbs:
          $ref: '#/components/schemas/Mpbs'
        letter:
          $ref: '#/components/schemas/FeeLetter'
        student_first_name:
          type: string
    FeeLetter:
      type: object
      properties:
        id:
          type: string
        approval_datetime:
          type: string
          format: date-time
        creation_datetime:
          type: string
          format: date-time
        ref:
          type: string
        status:
          $ref: '#/components/schemas/LetterStatus'
        file_url:
          type: string
...
    MpbsStatus:
      type: string
      enum:
        - PENDING
        - SUCCESS
        - FAILED
    MobileMoneyType:
      type: string
      enum:
        - MVOLA
        - ORANGE_MONEY
        - AIRTEL_MONEY
...
    CrupdateFeeTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/FeeTypeEnum'
        amount:
          type: integer
        number_of_payments:
          type: integer
    FeeTemplate:
      allOf:
        - $ref: '#/components/schemas/CrupdateFeeTemplate'
      type: object
      properties:
        creation_datetime:
          type: string
          format: date-time
...
    CrupdateMpbs:
      type: object
      properties:
        id:
          type: string
        student_id:
          type: string
        fee_id:
          type: string
        psp_id:
          type: string
          description: The couple `(fee_id, psp_id)` MUST be unique
        psp_type:
          $ref: '#/components/schemas/MobileMoneyType'
    Mpbs:
      allOf:
        - $ref: '#/components/schemas/CrupdateMpbs'
      type: object
      properties:
        amount:
          type: integer
        creation_datetime:
          type: string
          format: date-time
        psp_own_datetime_verification:
          type: string
          format: date-time
        successfully_verified_on:
          type: string
          format: date-time
        last_datetime_verification:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/MpbsStatus'
...
    FeeStatusEnum:
      type: string
      enum:
        - UNPAID
        - PAID
        - LATE
        - PENDING
    FeeTypeEnum:
      type: string
      enum:
        - TUITION
        - HARDWARE
        - STUDENT_INSURANCE
        - REMEDIAL_COSTS
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
      default: 15
...
    FeesStatistics:
      type: object
      properties:
        total_fees:
          type: integer
        paid_fees:
          type: integer
        unpaid_fees:
          type: integer
        late_fees:
          type: integer
        pending_transaction:
          type: integer
        paid_by_transaction:
          type: integer
        total_monthly_fees:
          type: integer
        total_yearly_fees:
          type: integer
...
    LetterStatus:
      type: string
      enum:
        - PENDING
        - RECEIVED
        - REJECTED
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error